feat: extract drawers to components
  * drawers are their own components
  * removed unused imports
